+++ src/process_inspector/unity/__init__.py
@@ -1,0 +1,5 @@
+from process_inspector.utils.importutils import get_platform_module
+
+platform_module = get_platform_module(__name__)
+
+UnityApp = platform_module.UnityApp